+++ descwl_shear_sims/sim.py
@@ -139,7 +139,8 @@ def make_sim(
         mock up a sky oversubtraction.  Default None.
     draw_method: string, optional
         Draw method for galaxy galsim objects, default 'auto'.  Set to
-        'phot' to get poisson noise.  Note this is much slower.
+        'phot' to get poisson noise, or use 'no_pixel' if the PSF
+        already has the  pixel in it.
     theta0: float, optional
         rotation angle of intrinsic galaxy shapes and positions on the sky,
         default 0, in units of radians
@@ -586,7 +587,12 @@ def make_exp(
         bright_info = []
 
     dm_wcs = make_dm_wcs(se_wcs)
-    dm_psf = make_dm_psf(psf=psf, psf_dim=psf_dim, wcs=se_wcs)
+    dm_psf = make_dm_psf(
+        psf=psf,
+        psf_dim=psf_dim,
+        wcs=se_wcs,
+        draw_method=draw_method if draw_method != 'phot' else 'auto',
+    )
 
     variance = image.copy()
     variance.array[:, :] = noise**2